backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
requests
python-dotenv
alembic
httpx
beautifulsoup4
lxml
asyncpg